refactor: manageModalSlice 파일 modal 디렉터리로 이동
