+++ src/components/Header.tsx
@@ -23,7 +23,7 @@ import {
   toggleArtist,
   toggleCategory,
   toggleArtistType,
-} from '../features/manageModalSlice';
+} from '../features/modal/manageModalSlice';
 import { useRouter } from '../hooks/useRouter';
 import { TextButton } from '../pages/Main/MainStyle';
 import media from '../utils/mediaQuery';
