+++ src/features/artists/modals/ArtistSelectModal.tsx
@@ -15,7 +15,7 @@ import {
   Artist,
   selectSelectedArtist,
 } from '../../events/services/setEventElemetsSlice';
-import { toggleEventArtist } from '../../manageModalSlice';
+import { toggleEventArtist } from '../../modal/manageModalSlice';
 import { useGetArtistsQuery } from '../services/artistsSlice';
 
 import {
