+++ src/features/categories/modals/CategorySelectModal.tsx
@@ -14,7 +14,7 @@ import {
   selectSelectedCategory,
   setCategory,
 } from '../../events/services/setEventElemetsSlice';
-import { toggleEventCategory } from '../../manageModalSlice';
+import { toggleEventCategory } from '../../modal/manageModalSlice';
 import {
   categoryType,
   useGetEventCategoryQuery,
