+++ src/pages/ManagePage/Artist/ArtistListItem.tsx
@@ -5,8 +5,8 @@ import deleteIcon from '../../../assets/icons/delete.svg';
 import editIcon from '../../../assets/icons/edit.svg';
 import { useDeleteArtistsMutation } from '../../../features/artists/services/artistsSlice';
 import { editArtistInfo } from '../../../features/artists/services/editArtistSlice';
-import { toggleEditArtist } from '../../../features/manageModalSlice';
-import { toggleEditGroup } from '../../../features/manageModalSlice';
+import { toggleEditArtist } from '../../../features/modal/manageModalSlice';
+import { toggleEditGroup } from '../../../features/modal/manageModalSlice';
 import { ArtistContent } from '../../../types/artistsType';
 
 import {
