+++ src/pages/ManagePage/Category/CategoryListItem.tsx
@@ -4,7 +4,7 @@ import deleteIcon from '../../../assets/icons/delete.svg';
 import editIcon from '../../../assets/icons/edit.svg';
 import { editCategoryInfo } from '../../../features/categories/services/editCategorySlice';
 import { useDeleteEventCategoryMutation } from '../../../features/categories/services/eventCategorySlice';
-import { toggleEditCategory } from '../../../features/manageModalSlice';
+import { toggleEditCategory } from '../../../features/modal/manageModalSlice';
 
 import * as S from './CategoryListItemStyle';
