+++ src/pages/MyPage/Bookmark/BookmarkFolders.tsx
@@ -5,7 +5,7 @@ import plusicon from '../../../assets/icons/cross.svg';
 import editicon from '../../../assets/icons/editpencil.svg';
 import starticon from '../../../assets/icons/starIcon.svg';
 import { useGetBookmarkFoldersQuery } from '../../../features/bookmarks/services/bookmarkFolderSlice';
-import { toggleAddBookmarkFolder } from '../../../features/manageModalSlice';
+import { toggleAddBookmarkFolder } from '../../../features/modal/manageModalSlice';
 import { BookmarkFolderType } from '../../../types/bookmarkFolderType';
 
 import BookmarkFolderItem from './BookmarkFolderItem';
